+++ lib/synchronization/condition_variable.dart
@@ -3,6 +3,7 @@ import 'dart:async';
 import 'package:meta/meta.dart';
 
 import '../src/synchronization/wait_queue.dart';
+import '../src/time_utils.dart';
 import 'lock.dart';
 
 /// A [ConditionVariable] is a synchronization primitive  that allows to wait
@@ -18,6 +19,10 @@ class ConditionVariable {
 
   ConditionVariable(this.lock);
 
+  /// Removes an element from the wait queue and waits for the lock to be
+  /// acquired.\
+  /// After the locks acquired, schedules the action specified in the removed
+  /// from the queue element to run.
   Future<void> notify() {
     if (_queue.isNotEmpty) {
       final awaiter = _queue.removeFirst();
@@ -29,6 +34,10 @@ class ConditionVariable {
     return _void;
   }
 
+  /// Removes all elements from the wait queue and waits for the lock to be
+  /// acquired for each element separately.\
+  /// After the locks acquired, schedules the actions specified in the removed
+  /// from the queue elements to run.
   Future<void> notifyAll() {
     while (_queue.isNotEmpty) {
       final awaiter = _queue.removeFirst();
@@ -40,16 +49,29 @@ class ConditionVariable {
     return _void;
   }
 
+  /// Releases the lock and waits for a notification.\
+  /// Upon receiving the notification, the lock will be reacquired. Accordingly,
+  /// upon exiting the method, the locked code will be entered.
+  ///
+  /// Returns `true` if the timeout has not expired; otherwise, returns `false`.
   @useResult
   Future<bool> tryWait(Duration timeout) async {
+    final started = TimeUtils.elapsedMicroseconds;
+    if (timeout.isNegative) {
+      throw ArgumentError.value(
+          timeout, 'timeout', 'Timeout must not be negative');
+    }
+
     await lock.release();
-    final waiter = _queue.enqueue(timeout);
-    return await waiter;
+    await _queue.enqueue();
+    return TimeUtils.elapsedMicroseconds - started < timeout.inMicroseconds;
   }
 
+  /// Releases the lock and waits for a notification.\
+  /// Upon receiving the notification, the lock will be reacquired. Accordingly,
+  /// upon exiting the method, the locked code will be entered.
   Future<void> wait() async {
     await lock.release();
-    final waiter = _queue.enqueue();
-    await waiter;
+    await _queue.enqueue();
   }
 }